+++ config/routes/annotations.yaml
@@ -1,16 +1,7 @@
 controllers:
     resource: '../../src/Controller/**/*'
-    exclude: 
-        - '../../src/Controller/Admin/'
-        - '../../src/Controller/HomeController.php'
     type: annotation
     name_prefix: ''
-
-admin:
-    resource: '../../src/Controller/Admin/'
-    type: annotation
-    prefix: '/admin'
-    name_prefix: 'admin_'
 
 kernel:
     resource: ../../src/Kernel.php
