+++ src/Controller/ActivityController.php
@@ -11,14 +11,34 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/activity', name: 'activity_')]
+#[Route('/activity')]
 class ActivityController extends AbstractController
 {
-    #[Route('/', name: 'index', methods: ['GET'])]
+    #[Route('/', name: 'activity_index', methods: ['GET'])]
     public function index(ActivityRepository $activityRepository): Response
     {
         return $this->render('activity/index.html.twig', [
             'activities' => $activityRepository->findAll(),
+        ]);
+    }
+
+    #[Route('/new', name: 'activity_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $activity = new Activity();
+        $form = $this->createForm(ActivityType::class, $activity);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($activity);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('activity_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('activity/new.html.twig', [
+            'activity' => $activity,
+            'form' => $form,
         ]);
     }
 
@@ -28,5 +48,34 @@ class ActivityController extends AbstractController
         return $this->render('activity/show.html.twig', [
             'activity' => $activity,
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'activity_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Activity $activity, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(ActivityType::class, $activity);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('activity_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('activity/edit.html.twig', [
+            'activity' => $activity,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'activity_delete', methods: ['POST'])]
+    public function delete(Request $request, Activity $activity, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$activity->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($activity);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('activity_index', [], Response::HTTP_SEE_OTHER);
     }
 }
