--- src/Controller/Admin/AdminController.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Controller\Admin;
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-
-#[Route('/', name: 'home_')]
-class AdminController extends AbstractController
-{
-    #[Route('/', name: 'index')]
-    public function index(): Response
-    {
-        return $this->render('admin/index.html.twig', [
-            'controller_name' => 'AdminController',
-        ]);
-    }
-    /**
-     * @Route("/logout", name="logout", methods={"GET"})
-     */
-    public function logout(): void
-    {
-    }
-}
