--- src/Controller/Admin/CountryController.php
@@ -1,73 +1,0 @@
-<?php
-
-namespace App\Controller\Admin;
-
-use App\Entity\Country;
-use App\Form\CountryType;
-use App\Repository\CountryRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-
-#[Route('/country', name: 'country_')]
-class CountryController extends AbstractController
-{
-    #[Route('/', name: 'index', methods: ['GET'])]
-    public function index(CountryRepository $countryRepository): Response
-    {
-        return $this->render('admin/country/index.html.twig', [
-            'countries' => $countryRepository->findAll(),
-        ]);
-    }
-
-    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
-    {
-        $country = new Country();
-        $form = $this->createForm(CountryType::class, $country);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($country);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('admin_country_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->renderForm('admin/country/new.html.twig', [
-            'country' => $country,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Country $country, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(CountryType::class, $country);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('admin/country_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->renderForm('admin/country/edit.html.twig', [
-            'country' => $country,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}', name: 'delete', methods: ['POST'])]
-    public function delete(Request $request, Country $country, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$country->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($country);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('admin/country_index', [], Response::HTTP_SEE_OTHER);
-    }
-}
