--- src/Controller/Admin/HotelController.php
@@ -1,73 +1,0 @@
-<?php
-
-namespace App\Controller\Admin;
-
-use App\Entity\Hotel;
-use App\Form\HotelType;
-use App\Repository\HotelRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-
-#[Route('/hotel', name: 'hotel_',)]
-class HotelController extends AbstractController
-{
-    #[Route('/', name: 'index', methods: ['GET'])]
-    public function index(HotelRepository $hotelRepository): Response
-    {
-        return $this->render('admin/hotel/index.html.twig', [
-            'hotels' => $hotelRepository->findAll(),
-        ]);
-    }
-
-    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
-    {
-        $hotel = new Hotel();
-        $form = $this->createForm(HotelType::class, $hotel);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($hotel);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('admin_hotel_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->renderForm('admin/hotel/new.html.twig', [
-            'hotel' => $hotel,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Hotel $hotel, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(HotelType::class, $hotel);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('hotel_index', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->renderForm('admin/hotel/edit.html.twig', [
-            'hotel' => $hotel,
-            'form' => $form,
-        ]);
-    }
-
-    #[Route('/{id}', name: 'delete', methods: ['POST'])]
-    public function delete(Request $request, Hotel $hotel, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$hotel->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($hotel);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('admin/hotel_index', [], Response::HTTP_SEE_OTHER);
-    }
-}
