+++ src/Controller/CityController.php
@@ -22,12 +22,60 @@ class CityController extends AbstractController
         ]);
     }
 
-  
+    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $city = new City();
+        $form = $this->createForm(CityType::class, $city);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($city);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('city_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('city/new.html.twig', [
+            'city' => $city,
+            'form' => $form,
+        ]);
+    }
+
     #[Route('/{id}', name: 'show', methods: ['GET'])]
     public function show(City $city): Response
     {
         return $this->render('city/show.html.twig', [
             'city' => $city,
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, City $city, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(CityType::class, $city);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('city_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('city/edit.html.twig', [
+            'city' => $city,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'delete', methods: ['POST'])]
+    public function delete(Request $request, City $city, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$city->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($city);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('city_index', [], Response::HTTP_SEE_OTHER);
     }
 }
