+++ src/Controller/CountryController.php
@@ -22,11 +22,60 @@ class CountryController extends AbstractController
         ]);
     }
 
+    #[Route('/new', name: 'new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $country = new Country();
+        $form = $this->createForm(CountryType::class, $country);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($country);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('country_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('country/new.html.twig', [
+            'country' => $country,
+            'form' => $form,
+        ]);
+    }
+
     #[Route('/{id}', name: 'show', methods: ['GET'])]
     public function show(Country $country): Response
     {
         return $this->render('country/show.html.twig', [
             'country' => $country,
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Country $country, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(CountryType::class, $country);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('country_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('country/edit.html.twig', [
+            'country' => $country,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'delete', methods: ['POST'])]
+    public function delete(Request $request, Country $country, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$country->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($country);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('country_index', [], Response::HTTP_SEE_OTHER);
     }
 }
