+++ src/Controller/HotelController.php
@@ -11,10 +11,10 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/hotel', name: 'hotel_')]
+#[Route('/hotel')]
 class HotelController extends AbstractController
 {
-    #[Route('/', name: 'index', methods: ['GET'])]
+    #[Route('/', name: 'hotel_index', methods: ['GET'])]
     public function index(HotelRepository $hotelRepository): Response
     {
         return $this->render('hotel/index.html.twig', [
@@ -22,11 +22,60 @@ class HotelController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'show', methods: ['GET'])]
+    #[Route('/new', name: 'hotel_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $hotel = new Hotel();
+        $form = $this->createForm(HotelType::class, $hotel);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($hotel);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('hotel_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('hotel/new.html.twig', [
+            'hotel' => $hotel,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'hotel_show', methods: ['GET'])]
     public function show(Hotel $hotel): Response
     {
         return $this->render('hotel/show.html.twig', [
             'hotel' => $hotel,
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'hotel_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Hotel $hotel, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(HotelType::class, $hotel);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('hotel_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('hotel/edit.html.twig', [
+            'hotel' => $hotel,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'hotel_delete', methods: ['POST'])]
+    public function delete(Request $request, Hotel $hotel, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$hotel->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($hotel);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('hotel_index', [], Response::HTTP_SEE_OTHER);
     }
 }
