+++ src/Controller/RestoController.php
@@ -11,10 +11,10 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/resto', name: 'resto_')]
+#[Route('/resto')]
 class RestoController extends AbstractController
 {
-    #[Route('/', name: 'index', methods: ['GET'])]
+    #[Route('/', name: 'resto_index', methods: ['GET'])]
     public function index(RestoRepository $restoRepository): Response
     {
         return $this->render('resto/index.html.twig', [
@@ -22,11 +22,60 @@ class RestoController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'show', methods: ['GET'])]
+    #[Route('/new', name: 'resto_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $resto = new Resto();
+        $form = $this->createForm(RestoType::class, $resto);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($resto);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('resto_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('resto/new.html.twig', [
+            'resto' => $resto,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'resto_show', methods: ['GET'])]
     public function show(Resto $resto): Response
     {
         return $this->render('resto/show.html.twig', [
             'resto' => $resto,
         ]);
+    }
+
+    #[Route('/{id}/edit', name: 'resto_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Resto $resto, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(RestoType::class, $resto);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('resto_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('resto/edit.html.twig', [
+            'resto' => $resto,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'resto_delete', methods: ['POST'])]
+    public function delete(Request $request, Resto $resto, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$resto->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($resto);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('resto_index', [], Response::HTTP_SEE_OTHER);
     }
 }
