--- src/DataFixtures/UserFixtures.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\DataFixtures;
-
-use App\Entity\User;
-use Doctrine\Persistence\ObjectManager;
-use Doctrine\Bundle\FixturesBundle\Fixture;
-use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
-
-class UserFixtures extends Fixture
-{
-    private UserPasswordHasherInterface $passwordHasher;
-
-    public function __construct(UserPasswordHasherInterface $passwordHasher)
-    {
-        $this->passwordHasher = $passwordHasher;
-    }
-
-    public function load(ObjectManager $manager): void
-    {
-        $user = new User();
-        $user->setEmail('admin@travelshaker.fr');
-        $user->setRoles(['ROLE_ADMIN']);
-        $user->setPassword($this->passwordHasher->hashPassword(
-            $user,
-            "travel007"
-        ));
-      
-        $manager->persist($user);
-        
-        $manager->flush();
-    }
-}
