--- templates/admin/_nav.html.twig
@@ -1,27 +1,0 @@
-<nav class="main-nav">
-    <div class="d-flex flex-column flex-shrink-0 p-3" style="width: 280px;">
-        <a href="/" class="d-flex align-items-center mb-3 mb-md-0 me-md-auto text-decoration-none">
-            <svg class="bi me-2" width="40" height="32">
-                <use xlink:href="#bootstrap"></use>
-            </svg>
-            <span class="fs-4">Travel Shaker</span>
-        </a>
-        <hr>
-        <ul class="nav nav-pills flex-column mb-auto">
-            {% set links = [
-                {'title': 'ACCUEIL', 'path': 'admin_home_index'},
-                {'title': 'MA CHECKLIST', 'path': 'admin_home_index'},
-                {'title': 'MES APPLIS VOYAGES', 'path': 'admin_home_index'},
-                {'title': 'DESTINATIONS', 'path': 'admin_home_index'},
-            ] %}
-            {% for link in links %}
-                <li class="nav-item">
-                    <a href="{{ path(link.path) }}"
-                       class="nav-link {{ app.request.pathInfo == path(link.path) ? 'active' : 'text-white' }}">
-                        {{ link.title }}
-                    </a>
-                </li>
-            {% endfor %}
-        </ul>
-    </div>
-</nav>
