--- templates/admin/country/index.html.twig
@@ -1,55 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Country index{% endblock %}
-
-{% block body %}
-    <h1>Country index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Name</th>
-                <th>Date</th>
-                <th>Duration</th>
-                <th>Hello</th>
-                <th>Thanku</th>
-                <th>Picture</th>
-                <th>Diving</th>
-                <th>Content</th>
-                <th>Slug</th>
-                <th>Links</th>
-                <th>Bye</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for country in countries %}
-            <tr>
-                <td>{{ country.id }}</td>
-                <td>{{ country.name }}</td>
-                <td>{{ country.date ? country.date|date('Y-m-d H:i:s') : '' }}</td>
-                <td>{{ country.duration }}</td>
-                <td>{{ country.hello }}</td>
-                <td>{{ country.thanku }}</td>
-                <td>{{ country.picture }}</td>
-                <td>{{ country.diving }}</td>
-                <td>{{ country.content }}</td>
-                <td>{{ country.slug }}</td>
-                <td>{{ country.links }}</td>
-                <td>{{ country.bye }}</td>
-                <td>
-                    <a href="{{ path('country_show', {'id': country.id}) }}">show</a>
-                    <a href="{{ path('country_edit', {'id': country.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="13">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('country_new') }}">Create new</a>
-{% endblock %}
