--- templates/admin/hotel/index.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Hotel index{% endblock %}
-
-{% block body %}
-    <h1>Hotel index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Name</th>
-                <th>Picture</th>
-                <th>Url</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for hotel in hotels %}
-            <tr>
-                <td>{{ hotel.id }}</td>
-                <td>{{ hotel.name }}</td>
-                <td>{{ hotel.picture }}</td>
-                <td>{{ hotel.url }}</td>
-                <td>
-                    <a href="{{ path('hotel_show', {'id': hotel.id}) }}">show</a>
-                    <a href="{{ path('hotel_edit', {'id': hotel.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="5">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('hotel_new') }}">Create new</a>
-{% endblock %}
