--- templates/admin/index.html copy.twig
@@ -1,45 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}City index{% endblock %}
-
-{% block body %}
-    <h1>City index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Name</th>
-                <th>Date</th>
-                <th>Picture</th>
-                <th>Duration</th>
-                <th>Content</th>
-                <th>Slug</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for city in cities %}
-            <tr>
-                <td>{{ city.id }}</td>
-                <td>{{ city.name }}</td>
-                <td>{{ city.date ? city.date|date('Y-m-d H:i:s') : '' }}</td>
-                <td>{{ city.picture }}</td>
-                <td>{{ city.duration }}</td>
-                <td>{{ city.content }}</td>
-                <td>{{ city.slug }}</td>
-                <td>
-                    <a href="{{ path('city_show', {'id': city.id}) }}">show</a>
-                    <a href="{{ path('city_edit', {'id': city.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="8">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('city_new') }}">Create new</a>
-{% endblock %}
