--- templates/admin/resto/index.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Resto index{% endblock %}
-
-{% block body %}
-    <h1>Resto index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Name</th>
-                <th>Picture</th>
-                <th>Url</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for resto in restos %}
-            <tr>
-                <td>{{ resto.id }}</td>
-                <td>{{ resto.name }}</td>
-                <td>{{ resto.picture }}</td>
-                <td>{{ resto.url }}</td>
-                <td>
-                    <a href="{{ path('resto_show', {'id': resto.id}) }}">show</a>
-                    <a href="{{ path('resto_edit', {'id': resto.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="5">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('resto_new') }}">Create new</a>
-{% endblock %}
