+++ templates/hotel/_delete_form.html.twig
