--- templates/login/index.html.twig
@@ -1,29 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Login{% endblock %}
-
-{% block body %}
-{% if error %}
-        <div>{{ error.messageKey|trans(error.messageData, 'security') }}</div>
-    {% endif %}
-    <div class="d-flex justify-content-center">
-        <form style="margin-top: 200px" action="{{ path('login') }}" method="post">
-            <div class="md-3">
-                <label class="form-label" for="username">Username:</label>
-                <input class="form-control" type="text" id="username" name="_username" value="{{ last_username }}"/>
-            </div>
-            <div class="md-3 mt-2">
-                <label class="form-label" for="password">{{ 'Mot de passe'|trans }}</label>
-                <input class="form-control" type="password" id="password" name="_password"/>
-            </div>
-        <div>
-            {# If you want to control the URL the user is redirected to on success#}
-            <input type="hidden" name="_target_path" value="{{ path('admin_home_index') }}"/>
-            <input type="hidden" name="_csrf_token" value="{{ csrf_token('authenticate') }}">
-    <div class="d-flex justify-content-end mt-3 mb-2">
-            <button class="btn btn-primary border" type="submit">{{ 'Se connecter'|trans }}</button>
-        </div>
-        </div>
-        </form>
-    </div>
-{% endblock %}
