+++ templates/resto/_delete_form.html.twig
